# ApexWorld_Backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Apex World REPMS API",
    "description": "Backend API for Apex World REPMS",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/BuyerBooking/{id}/reschedule": {
      "patch": {
        "tags": [
          "Buyer - Site-Visit Booking"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/BuyerBooking/book": {
      "post": {
        "tags": [
          "Buyer - Site-Visit Booking"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique key to ensure the request is processed exactly once (e.g., a GUID).",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/Property": {
      "get": {
        "tags": [
          "Buyer & Public - Properties"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Property/{id}": {
      "get": {
        "tags": [
          "Buyer & Public - Properties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Property/search": {
      "get": {
        "tags": [
          "Buyer & Public - Properties"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "Approved"
            }
          },
          {
            "name": "name",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BookingRequestDto": {
        "type": "object",
        "properties": {
          "propertyId": {
            "type": "integer",
            "format": "int32"
          },
          "buyerId": {
            "type": "string",
            "nullable": true
          },
          "offerAmount": {
            "type": "number",
            "format": "double"
          },
          "scheduledDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "permanentAddress": {
            "type": "string",
            "nullable": true
          },
          "purposeOfPurchase": {
            "type": "string",
            "nullable": true
          },
          "timeSlot": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RescheduleRequest": {
        "type": "object",
        "properties": {
          "newDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 1safsfsdfdfd\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}